player: big clean-up of includes and platform-specifics
more clean-ups to follow
