+++ amiga/Makefile.vbcc
@@ -35,7 +35,7 @@ endif
 
 # Objects
 LIB_OBJ= wm_error.o file_io.o lock.o wildmidi_lib.o reverb.o gus_pat.o f_xmidi.o f_mus.o f_hmp.o f_midi.o f_hmi.o mus2mid.o xmi2mid.o internal_midi.o patches.o sample.o
-PLAYER_OBJ= getopt_long.o wm_tty.o out_wave.o out_ahi.o fileops.o amiga.o wildmidi.o
+PLAYER_OBJ= amiga.o wm_tty.o msleep.o getopt_long.o out_wave.o out_ahi.o wildmidi.o
 
 # Build targets
 .PHONY: clean distclean
