+++ djgpp/Makefile
@@ -60,7 +60,7 @@ endif
 
 # Objects
 LIB_OBJ= wm_error.o file_io.o lock.o wildmidi_lib.o reverb.o gus_pat.o f_xmidi.o f_mus.o f_hmp.o f_midi.o f_hmi.o mus2mid.o xmi2mid.o internal_midi.o patches.o sample.o
-PLAYER_OBJ= $(SB_OBJ) fileops.o getopt_long.o out_dossb.o out_wave.o wm_tty.o wildmidi.o
+PLAYER_OBJ= wm_tty.o msleep.o getopt_long.o $(SB_OBJ) out_dossb.o out_wave.o wildmidi.o
 
 # Build targets
 TARGETS = libWildMidi.a wildmidi.exe libWildMidi_dxe.a wildmidi.dxe
