+++ mingw/Makefile
@@ -53,7 +53,7 @@ LDLIBS_EXE+=-L. -l$(LIBNAME)
 # Objects
 LIB_OBJ = wm_error.o file_io.o lock.o wildmidi_lib.o reverb.o gus_pat.o
 LIB_OBJ+= f_xmidi.o f_mus.o f_hmp.o f_midi.o f_hmi.o mus2mid.o xmi2mid.o internal_midi.o patches.o sample.o
-PLAYER_OBJ = wm_tty.o getopt_long.o wildmidi.o out_wave.o out_win32mm.o fileops.o
+PLAYER_OBJ = wm_tty.o msleep.o getopt_long.o out_wave.o out_win32mm.o wildmidi.o
 # out_openal.o
 
 .PHONY: clean distclean
@@ -98,7 +98,7 @@ out_win32mm.o: ../src/player/out_win32mm.c
 	$(CC) -c $(CFLAGS_EXE) -o $@ $<
 out_openal.o: ../src/player/out_openal.c
 	$(CC) -c $(CFLAGS_EXE) -o $@ $<
-fileops.o: ../src/player/fileops.c
+msleep.o: ../src/player/msleep.c
 	$(CC) -c $(CFLAGS_EXE) -o $@ $<
 wm_tty.o: ../src/player/wm_tty.c
 	$(CC) -c $(CFLAGS_EXE) -o $@ $<
