+++ os2/makefile.emx
@@ -29,8 +29,7 @@ CFLAGS_LIB= $(CFLAGS) -DWILDMIDI_BUILD
 CFLAGS_EXE= $(CFLAGS)
 
 OBJ=wm_error.o file_io.o lock.o wildmidi_lib.o reverb.o gus_pat.o f_xmidi.o f_mus.o f_hmp.o f_midi.o f_hmi.o mus2mid.o xmi2mid.o internal_midi.o patches.o sample.o
-PLAYER_OBJ=wildmidi.o getopt_long.o out_wave.o out_dart.o fileops.o
-# wm_tty.o
+PLAYER_OBJ=wm_tty.o msleep.o getopt_long.o out_wave.o out_dart.o wildmidi.o
 
 all: $(LIBSTATIC) $(PLAYER_STATIC)
 
@@ -47,7 +46,7 @@ $(PLAYER_STATIC): $(LIBSTATIC) $(PLAYER_OBJ)
 # rules for player objs:
 getopt_long.o: ../src/getopt_long.c
 	$(CC) -c $(CFLAGS_EXE) -o $@ $<
-fileops.o: ../src/player/fileops.c
+msleep.o: ../src/player/msleep.c
 	$(CC) -c $(CFLAGS_EXE) -o $@ $<
 out_wave.o: ../src/player/out_wave.c
 	$(CC) -c $(CFLAGS_EXE) -o $@ $<
